# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alexander Evan API",
        "description": "Dokumentasi API Web Untuk Belajar",
        "contact": {
            "name": "Energeek",
            "email": "energeekmail@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Base URL"
        }
    ],
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "API|MASTER|INDEX JOBS"
                ],
                "summary": "Jobs index",
                "operationId": "6b307032ccc7685af33c78dee9152c22",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "$ref": "#/components/responses/globalResponse"
                    }
                }
            }
        },
        "/api/jobs/store/": {
            "post": {
                "tags": [
                    "Api|Master|STORE JOBS"
                ],
                "summary": "Jobs store",
                "operationId": "5a583604437d49b35a00faff5440606b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "$ref": "#/components/responses/globalResponse"
                    }
                }
            }
        },
        "/api/jobs/show/{id}": {
            "get": {
                "tags": [
                    "API|MASTER|SHOW JOBS"
                ],
                "summary": "Jobs show",
                "operationId": "7016f88c2a8a323115fa99efb6b99e94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "$ref": "#/components/responses/globalResponse"
                    }
                }
            }
        },
        "/api/jobs/update/{id}": {
            "put": {
                "tags": [
                    "Api|MASTER|UPDATE JOBS"
                ],
                "summary": "Candidates update",
                "operationId": "9af8dbd3f300ab2882cdc7131cce51b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "$ref": "#/components/responses/globalResponse"
                    }
                }
            }
        },
        "/api/jobs/delete/{id}": {
            "delete": {
                "tags": [
                    "Api|Master|DELETE JOBS"
                ],
                "summary": "Jobs delete-file",
                "operationId": "ec52ac78de27234d4981dacb0533669a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "$ref": "#/components/responses/globalResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "responseFormatter": {
                "properties": {
                    "status": {
                        "properties": {
                            "code": {
                                "type": "integer"
                            },
                            "message": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "object",
                        "default": null
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "globalResponse": {
                "description": "OK",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/responseFormatter"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "authBearerToken": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "authBearerToken": []
        }
    ]
}